+++ quartz/src/commonMain/kotlin/com/vitorpamplona/quartz/nip47WalletConnect/kotlinSerialization/Nip47RequestKSerializer.kt
@@ -20,32 +20,32 @@
  */
 package com.vitorpamplona.quartz.nip47WalletConnect.kotlinSerialization
 
-import com.vitorpamplona.quartz.nip47WalletConnect.CancelHoldInvoiceMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.CancelHoldInvoiceParams
-import com.vitorpamplona.quartz.nip47WalletConnect.CreateConnectionMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.CreateConnectionParams
-import com.vitorpamplona.quartz.nip47WalletConnect.GetBalanceMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.GetBudgetMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.GetInfoMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.ListTransactionsMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.ListTransactionsParams
-import com.vitorpamplona.quartz.nip47WalletConnect.LookupInvoiceMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.LookupInvoiceParams
-import com.vitorpamplona.quartz.nip47WalletConnect.MakeHoldInvoiceMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.MakeHoldInvoiceParams
-import com.vitorpamplona.quartz.nip47WalletConnect.MakeInvoiceMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.MakeInvoiceParams
-import com.vitorpamplona.quartz.nip47WalletConnect.NwcMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.PayInvoiceMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.PayInvoiceParams
-import com.vitorpamplona.quartz.nip47WalletConnect.PayKeysendMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.PayKeysendParams
-import com.vitorpamplona.quartz.nip47WalletConnect.Request
-import com.vitorpamplona.quartz.nip47WalletConnect.SettleHoldInvoiceMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.SettleHoldInvoiceParams
-import com.vitorpamplona.quartz.nip47WalletConnect.SignMessageMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.SignMessageParams
-import com.vitorpamplona.quartz.nip47WalletConnect.TlvRecord
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.CancelHoldInvoiceMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.CancelHoldInvoiceParams
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.CreateConnectionMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.CreateConnectionParams
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.GetBalanceMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.GetBudgetMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.GetInfoMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.ListTransactionsMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.ListTransactionsParams
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.LookupInvoiceMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.LookupInvoiceParams
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.MakeHoldInvoiceMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.MakeHoldInvoiceParams
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.MakeInvoiceMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.MakeInvoiceParams
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.NwcMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.PayInvoiceMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.PayInvoiceParams
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.PayKeysendMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.PayKeysendParams
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.Request
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.SettleHoldInvoiceMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.SettleHoldInvoiceParams
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.SignMessageMethod
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.SignMessageParams
+import com.vitorpamplona.quartz.nip47WalletConnect.rpc.TlvRecord
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.descriptors.buildClassSerialDescriptor
